feat : implement initial rendering loop for editor
