--- Includes/VoxFlow/Editor/VoxEngine.hpp
@@ -1,22 +1,0 @@
-// Author : snowapril
-
-#ifndef VOXEL_FLOW_VOX_ENGINE_HPP
-#define VOXEL_FLOW_VOX_ENGINE_HPP
-
-namespace VoxFlow
-{
-class RenderDevice;
-
-class VoxEngine
-{
- public:
-    VoxEngine();
-    ~VoxEngine();
-
- protected:
- private:
-    RenderDevice* _renderDevice = nullptr;
-};
-}  // namespace VoxFlow
-
-#endif
