+++ Sources/VoxFlow/Editor/VoxEditor.cpp
@@ -1,22 +1,68 @@
 // Author : snowapril
 
-#include <glfw/glfw3.h>
-#include <VoxFlow/Editor/VoxEditor.hpp>
-#include <VoxFlow/Editor/VoxEngine.hpp>
 #include <chrono>
+#include <VoxFlow/Editor/VoxEditor.hpp>
+#include <VoxFlow/Core/RenderDevice.hpp>
+#include <VoxFlow/Core/Devices/LogicalDevice.hpp>
+#include <GLFW/glfw3.h>
 
 namespace VoxFlow
 {
 
 VoxEditor::VoxEditor()
 {
-    _voxEngine = new VoxEngine();
+    if (glfwInit() == GLFW_FALSE)
+    {
+        const char* glfwErrorMsg = nullptr;
+        glfwGetError(&glfwErrorMsg);
+        VOX_ASSERT(false, "Failed to initialize GLWF. LastError : %s",
+                   glfwErrorMsg);
+        return;
+    }
+
+    glfwSetErrorCallback(DebugUtil::GlfwDebugCallback);
+
+    if (glfwVulkanSupported() == GLFW_FALSE)
+    {
+        const char* glfwErrorMsg = nullptr;
+        glfwGetError(&glfwErrorMsg);
+        VOX_ASSERT(false,
+                   "Current availiability of loader and ICD does not support "
+                   "Vulkan. LastError : %s",
+                   glfwErrorMsg);
+        return;
+    }
+
+    constexpr bool enableDebugLayer = true;
+
+    Context context(enableDebugLayer);
+    context.setVersion(1, 3);
+
+    // Add GLFW required instace extensions and swapchain device extension
+    {
+        uint32_t numGlfwExtensions = 0;
+        const char** glfwExtensions = nullptr;
+        glfwExtensions = glfwGetRequiredInstanceExtensions(&numGlfwExtensions);
+        for (uint32_t i = 0; i < numGlfwExtensions; ++i)
+        {
+            context.addInstanceExtension(glfwExtensions[i]);
+        }
+        context.addDeviceExtension(VK_KHR_SWAPCHAIN_EXTENSION_NAME);
+    }
+    context.addDeviceExtension(VK_KHR_TIMELINE_SEMAPHORE_EXTENSION_NAME);
+
+    _renderDevice = new RenderDevice(context);
+    _renderDevice->getLogicalDevice(0)->addSwapChain("VoxFlow Editor", glm::ivec2(1280, 920));
 }
 
 VoxEditor::~VoxEditor()
 {
-    if (_voxEngine != nullptr)
-        delete _voxEngine;
+    if (_renderDevice != nullptr)
+    {
+        delete _renderDevice;
+    }
+
+    glfwTerminate();
 }
 
 void VoxEditor::runEditorLoop()
@@ -57,6 +103,9 @@ void VoxEditor::updateFrame()
 
 void VoxEditor::renderFrame()
 {
+    // TODO(snowapril) : customize logical device index
+    _renderDevice->beginFrame(0);
+    _renderDevice->presentSwapChains(0);
 }
 
 void VoxEditor::postRenderFrame()
